+++ packages/dice/src/guards/isD.ts
@@ -1,5 +1,5 @@
-import { D } from '~src/D'
-import type { BaseD } from '~types'
+import { D } from "../D";
+import type { BaseD } from "../types";
 
 export function isD(arg: unknown): arg is BaseD<string[] | number> {
   return arg instanceof D
