+++ packages/dice/src/index.ts
@@ -1,0 +1,4 @@
+export { D } from './D'
+export * from './premadeDice'
+export { roll } from './roll'
+export type * from './types'
