+++ packages/dice/src/utils/calculateTotal.ts
