+++ packages/dice/src/utils/coreRandom.ts
