+++ packages/dice/src/utils/coreSpreadRolls.ts
