+++ packages/dice/src/utils/generateNumericalFaces.ts
