+++ packages/notation/src/validateNotation.ts
@@ -1,7 +1,7 @@
+import { optionsConverter } from "@randsum/core"
 import { isDiceNotation } from "./isDiceNotation"
 import type { CustomDiceNotation, CustomValidationResult, InvalidValidationResult, NumericDiceNotation, NumericValidationResult, ValidationResult } from "./types"
 import { notationToOptions } from "./utils/notationToOptions"
-import { optionsConverter } from "./utils/optionsConverter"
 
 /**
  * Validates dice notation strings and provides detailed information about their structure.
