--- src/guards/isDiceNotationArg.ts
@@ -1,12 +1,0 @@
-import { completeRollPattern, coreNotationPattern } from '~patterns'
-import type { DiceNotation } from '~types'
-
-export function isDiceNotationArg(argument: unknown): argument is DiceNotation {
-  const notAString = typeof argument !== 'string'
-  const basicTest = !!coreNotationPattern.test(String(argument))
-  if (!basicTest || notAString) return false
-
-  const cleanArg = argument.replace(/\s/g, '')
-
-  return cleanArg.replace(completeRollPattern, '').length === 0
-}
