--- src/index.ts
@@ -1,5 +1,0 @@
-export { D } from '~src/D'
-export * from '~src/premadeDice'
-export { roll } from '~src/roll'
-export type * from '~src/types'
-export { validateNotation } from '~src/validateNotation'
