--- src/utils/extractMatches.ts
@@ -1,8 +1,0 @@
-export function extractMatches(
-  notationString: string,
-  pattern: RegExp
-): string[] {
-  return [...(notationString.matchAll(pattern) || [])].map(
-    (matches) => matches[0]
-  )
-}
